add: admin CRUD

+++ app/Http/Controllers/Api/ItemController.php
@@ -44,7 +44,7 @@ class ItemController extends Controller
             Image::create([
                 'item_id' => $item->id,
                 'item_image' => $imagePath,
-                'item_image_order' => $request->input('images.'.$index.'.item_image_order'), // Corrected line
+                'item_image_order' => $request->input('images.'.$index.'.item_image_order'),
             ]);
         }
         
@@ -67,50 +67,44 @@ class ItemController extends Controller
      */
     public function update(UpdateItemRequest $request, Item $item)
     {
+        $existingImages = $item->images->pluck('item_image')->toArray();
+        
         $data = $request->validated();
+        
+        $item->update($data);
+        
+        $categoryIds = $request->input('categories');
+        $item->categories()->sync($categoryIds);
     
-        if (isset($data['images']) && is_array($data['images'])) {
-            foreach ($item->images as $image) {
-                Storage::disk('public')->delete($image->item_image);
-                $image->delete();
+        if ($request->file('images')){
+            $images = $request->file('images');
+            foreach ($images as $index => $imageData) {
+                $imagePath = $imageData['item_image']->store('images', 'public');
+    
+                Image::create([
+                    'item_id' => $item->id,
+                    'item_image' => $imagePath,
+                    'item_image_order' => $request->input('images.'.$index.'.item_image_order'),                ]);
             }
-
-            foreach ($data['images'] as $imageData) {
-                // Check if $imageData is an instance of UploadedFile
-                if ($imageData instanceof UploadedFile) {
-                    // Handle file upload logic
-                    $imagePath = $imageData['item_image']->store('images', 'public');
-            
-                    $imageId = $imageData['image_id']; // Assuming you have an 'image_id' key in your image data
-            
-                    // Delete previous image with the same ID
-                    Image::find($imageId)->delete();
-            
-                    // Create a new record for the updated image
-                    Image::create([
-                        'item_id' => $item->id,
-                        'item_image' => $imagePath,
-                        'item_image_order' => $imageData['item_image_order'],
-                    ]);
-                }
-            }            
+        } else {
+            $imagesData = $request->input('images');
+            foreach ($imagesData as $index => $imageData) {
+                $imageOrder = $imageData['item_image_order'];
+    
+                Image::where('item_id', $item->id)
+                    ->where('item_image', $imageData['item_image'])
+                    ->update(['item_image_order' => $imageOrder]);
+            }
         }
+
+        $deletedImages = array_diff($existingImages, $request->input('images.*.item_image'));
+        foreach ($deletedImages as $deletedImage) {
+            Storage::disk('public')->delete($deletedImage);
     
-        // Update other item fields
-        $item->update([
-            'item_name' => $data['item_name'],
-            'item_price' => $data['item_price'],
-            'item_description' => $data['item_description'],
-            'tokopedia_link' => $data['tokopedia_link'],
-            'shoppee_link' => $data['shoppee_link'],
-            'whatsapp_link' => $data['whatsapp_link'],
-            'available_stock' => $data['available_stock'],
-        ]);
-    
-        // Handle categories
-        $item->categories()->sync($data['categories']);
-    
-        return response(new ItemResource($item), 200);
+            Image::where('item_id', $item->id)->where('item_image', $deletedImage)->delete();
+        }
+        
+        return new ItemResource($item);
     }
     
     /**
@@ -118,21 +112,17 @@ class ItemController extends Controller
      */
     public function destroy(Item $item)
     {
-        $images = $item->images;
-        $categories = $item->categories;
-    
-        foreach ($images as $image) {
-            Storage::delete($image->item_image);
+        // Check if there are images associated with the item
+        if ($item->images->isNotEmpty()) {
+            foreach ($item->images as $image) {
+                Storage::delete($image->item_image);
+            }
+
+            $item->images()->delete();
         }
-    
-        $item->images()->delete();
-    
-        $item->categories()->detach();
-    
-        foreach ($categories as $category) {
-            $category->items()->detach($item->id);
-        }
-    
+
+
+        // Delete the item
         $item->delete();
     }
 }
